+++ packages/client-node/__tests__/utils/http_stubs.ts
@@ -134,6 +134,6 @@ export class MyTestHttpConnection extends NodeBaseConnection {
     return {} as any
   }
   public getDefaultHeaders() {
-    return this.headers
+    return this.buildRequestHeaders()
   }
 }